Move `quickcheck-laws` out of dependencies
I presume `quickcheck-laws` is just a devDependency. This PR is to move it to that section of the `bower.json` file.

Mostly this is for correctness, but it's also because I'm being hit by a bug where documentation doesn't generate due to something around the Quickcheck library, and BigInt is one of the libs I have installed. While diagnosing I noticed that the laws package was incorrectly categorized.